Update HyperloopControlV0.py
changed while loop condition. added comments about what still needs to be done. arduino file needs to be actually imported, system check is not at all done, for failsafes we need velocity and position, need to finish nested loops in failsafe, need to change halt() function in ESC file so that it actually turns ON brakes in addition to what it already does, which is just turning OFF the motors.